+++ include/simpleble/Exceptions.h
@@ -48,6 +48,11 @@ class OperationFailed : public BaseException {
     OperationFailed();
 };
 
+class WinRTException : public BaseException {
+  public:
+    WinRTException(int32_t err_code, const std::string &err_msg);
+};
+
 }  // namespace Exception
 
 }  // namespace SimpleBLE